+++ tests/testthat/test-002-model.R
@@ -3,9 +3,9 @@
 farms_count <- 10
 fields_max <- 5
 
-dataset.train <- apus::createApusDataset(farms = NULL, cultivations = apus::cultivations, fertilizers = apus::fertilizers, fields_max = fields_max, device = 'cpu')
+dataset.train <- apus::createApusDataset(farms = NULL, cultivations = apus::cultivations, fertilizers = apus::fertilizers, fines = apus::fines, fields_max = fields_max, device = 'cpu')
 farms.valid <-  apus:::createSyntheticFarms(farms_count = farms_count, fields_max = fields_max)
-dataset.valid<- apus::createApusDataset(farms = farms.valid, cultivations = apus::cultivations, fertilizers = apus::fertilizers, fields_max = fields_max, device = 'cpu')
+dataset.valid <- apus::createApusDataset(farms = farms.valid, cultivations = apus::cultivations, fertilizers = apus::fertilizers, fines = apus::fines, fields_max = fields_max, device = 'cpu')
 
 model <- apus::createApusModel(dataset.train, dataset.valid, device = 'cpu', epochs = 2)
 dl <- torch::dataloader(dataset.train, batch_size = farms_count)
@@ -14,6 +14,7 @@ batch <- dl$.iter()
 batch <- batch$.next()
 fields <- batch$fields
 fertilizers <- batch$fertilizers
+fines <- batch$fines
 doses <- model(fields, fertilizers)
 
 test_that("Create model and run a forward pass", {
@@ -42,9 +43,14 @@ test_that("Calculate revenue for module 4: Revenue of harvested crops", {
   expect_length(as.numeric(module4), farms_count)
 })
 
+module5 <- calculatePenaltyModule5(doses, fields, fertilizers, fines)
 
+test_that("Calculate penalties for module 5: Penalties for exceeding legal limits", {
+  expect_contains(class(module5), 'torch_tensor')
+  expect_length(as.numeric(module5), farms_count)
+})
 
-cost <- calculateCost(doses, fields, fertilizers)
+cost <- calculateCost(doses, fields, fertilizers, fines)
 
 test_that("Calculate overall cost", {
   expect_contains(class(cost), 'torch_tensor')