+++ NEWS.md
@@ -2,5 +2,5 @@
 
 ## Added
 * Adds `apus` object with the functions `addField`, `trainModel` and `optimizeFertilizerChoice`
-* Adds for the cost function module 1 and 2
+* Adds for the cost function module 1,4 and 5
 * Adds default table for `cultivations`, `fertilizers`, `parameters` and `fines`